% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeFunctions.R
\name{linkNames}
\alias{linkNames}
\title{Link Names of Students and their Study to Registration Numbers}
\usage{
linkNames(regFile, responses)
}
\description{
Link Names of Students and their Study to Registration Numbers
}
\examples{
\dontrun{
responses <- linkNames(regFile, responses)}
}
