% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadKey.R
\name{loadKey}
\alias{loadKey}
\title{Load the answer key file for grading.}
\usage{
loadKey(keyFile)
}
\arguments{
\item{keyFile}{A character string specifying the location and name of the csv
or (x)lsx file containing the answer key to the exam items (questions).}
}
\value{
A data.frame object with in the first column the exam version and
   the following columns giving the answers to the exam items (questions).
}
\description{
\code{loadKey()} loads a file with the exam answer key for all versions. This
file should be a comma or semicolon separated value file (csv) or a Microsoft
Excel file (xls or xlsx) with on its rows the exam versions (e.g. "A", "B",
"C", "D") and on its columns the correct responses to the multiple choice
test items (questions). The first column specifies the exam version (e.g.
"A", "B", "C", "D").
}
\examples{
\dontrun{
key <- loadKey(keyFile)}
}
