% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{loadResponses}
\alias{loadResponses}
\title{Load Student Responses to Exam Test Items (Questions).}
\usage{
loadResponses(respFile, noItems = ncol(key) - 1)
}
\arguments{
\item{respFile}{A character string specifying the location and name of
the csv or (x)lsx file containing the student responses to the multiple
choice exam test items (questions).}

\item{noItems}{An integer, not exceeding 40, specifying the number of multiple
choice exam test items (questions) on the answers sheet to be considered.
If a \code{key} object has been created with the \code{\link{loadKey}}
function prior to using the \code{loadResponses} function the \code{noItems}
function argument repuires no specification and will created automatically.}
}
\value{
A data.frame object with in the first column the student registration
   numbers of students taking the exam. The second column specifies the exam
   version the student filled. Subsequent columns give the filled in answers
   to the multiple choice exam test items (questions) as filled by the
   students, with a maximum of 40 test items (questions).
}
\description{
\code{loadResponses()} loads a file with the student responses to the multi-
ple choice exam test items (questions). This file should be a FormScanner
generated comma or semicolon separated value csv-file or a Microsoft Excel
(x)lsx-file generated by Wageningen University & Research ICT/Edusuport, or
by means of the \code{\link{formScanner2Excel}} function. The file has on each
row: the student registration number, exam version, and the responses to the
multiple choice exam test items (questions).
}
\examples{
\dontrun{
responses <- loadResponses(respFile)}
}
