% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMC.R
\name{processMC}
\alias{processMC}
\title{GUI: Conversion of FromScanner File into Microsoft Excel File.}
\usage{
processMC()
}
\value{
A Microsoft Excel workbook xlsx-file with the name specified in one
  of the GUI questions.
}
\description{
Graphical User Interface guided conversion of a FormScanner csv-file into a
Microsoft Excel xlsx-file.
}
\seealso{
\code{\link{formScanner2Excel}}
}
