% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wurmc-package.R
\docType{package}
\name{wurmc}
\alias{wurmc}
\title{wurmc}
\description{
Package for processing and grading multiple choice exams at Wageningen
University and Research.
}
\details{
Multiple choice answers sheets from students at Wageningen University and
Research are generally scanned and processed into a Microsoft Excel worbook,
containing a worksheet named with the course code, by EDUsupport (part of
ITsupport). The used multiple choice answer sheets, however, also allow for
scanning and processing with the free, open source OMR (optical mark
recognition) software FormScanner (\url{http://www.formscanner.org/}). This
package can convert the resulting FormScanner csv-file (in reality a
semicolon separated file) into a Microsoft Excel workbook (xlsx-file) with
the same content layout as generated by EDUsupport.

Next the Microsoft Excel workbook is used as one of the required inputs for
grading the student multiple choice answer sheets. Other required files are:
\itemize{
 \item a file containing the answer key for the various versions (at most 4:
 A, B, C, D) used,
 \item a file containing the students that registered for the exam,
 \item a file containing the practicum information about the students,
 whether they passed the practicum and whether or not they have earned a
 bonus to be added to the exam grade.
}
}
