% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeFunctions.R
\name{gradeExam}
\alias{gradeExam}
\title{Grade the Multiple Choice Exams with Correction for Guessing}
\usage{
gradeExam(
  responses,
  noItems = length(grep("^I", colnames(responses))),
  noOptions = 4
)
}
\arguments{
\item{responses}{A data.frame object, as created with the
\code{\link{loadResponses}} function modified by the
\code{\link{numberCorrect}} and \code{\link{linkNames}} functions.}

\item{noItems}{An integer, not exceeding 40, specifying the number of multiple
choice exam test items (questions) on the answers sheet to be considered.
If a \code{key} object has been created with the \code{\link{loadKey}}
function prior to using the \code{loadResponses} function the \code{noItems}
function argument repuires no specification and will created automatically.}

\item{noOptions}{An integer, which specifies the number of options for each
multiple choice exam test item (question). The default is 4 options ("A",
"B", "C", and "D").}
}
\value{
The \code{responses} data.frame object extended with a column giving
   the calculated exam grade (Exam grade) for each student.
}
\description{
\code{gradeExam()} calculates the exam grades achieved based on the responses
to the multiple choice exam test items (questions). The exam grade is
corrected for guessing.
}
\examples{
\dontrun{
responses <- gradeExam(responses, noItems, noOptions)}
}
\seealso{
\code{\link{loadResponses}}, \code{\link{loadKey}},
   \code{\link{numberCorrect}}, \code{\link{linkNames}}
}
