% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeFunctions.R
\name{linkNames}
\alias{linkNames}
\title{Link Names of Students and their Study to Registration Numbers.}
\usage{
linkNames(regFile, responses)
}
\arguments{
\item{regFile}{A character string specifying the location and name of
the csv or (x)lsx file containing the student's registration number, name
and study for all students, which registered for the exam via the Student
Service Centre (SSC).}

\item{responses}{A data.frame object, as created with the
\code{\link{loadResponses}} function and modified by the
\code{\link{numberCorrect}} function.}
}
\value{
The input \code{responses} data.frame object extended with three columns
   indicating the corrected registration number (Corrected reg. number),
   Name, and Study of all students.
}
\description{
\code{linkNames()} compares the student registration numbers scanned from
the multiple choice answer sheets (filled by students during the exam) with
the student registration numbers from the list of students that registered
for the exam via the Student Service Centre (SSC). In case the registration
number was incorrectly filled on the answer sheet this function provides the
opportunity to manually correct this. From the exam registration list the
student's name and study are matched with the student registration number and
added to the responses object.
}
\examples{
\dontrun{
responses <- linkNames(regFile, responses)}
}
\seealso{
\code{\link{loadResponses}}, \code{\link{numberCorrect}}
}
