% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{loadResponses}
\alias{loadResponses}
\title{Load Student Responses File}
\usage{
loadResponses(respFile, noItems = ncol(key) - 1)
}
\arguments{
\item{respFile}{A character string (\strong{REQUIRED}) specifying the location
and name of the \code{.csv} or \code{.xls(x)} file containing the student responses
to the multiple choice exam test items (questions).}

\item{noItems}{An integer (\strong{OPTIONAL}) specifying the number of multiple
choice exam test items (questions) on the answers sheet to be considered
and matching the number of test items (questions) in the answer key file.
If a \code{key} object has been created with the \code{\link{loadKey}()}
function prior to using the \code{loadResponses()} function the \code{noItems}
function argument requires no specification and will be filled
automatically.}
}
\value{
A data.frame object with in the first column the student registration
numbers of students taking the exam. The second column specifies the exam
version the student filled. Subsequent columns give the filled in answers
to the multiple choice exam test items (questions) as filled by the
students, with a maximum of 40 test items (questions).
}
\description{
Loads a file containing the student responses to the multiple choice exam
test items (questions).
}
\details{
The \code{loadResponses()} function loads a file with the student responses to the
multiple choice exam test items (questions). This file should be a FormScanner
(\url{http://www.formscanner.org/}) generated comma or semicolon separated
value \code{.csv} file or a Microsoft Excel \code{.xls(x)} file generated by Wageningen
University & Research EDUsuport, or by means of the
\code{\link{formScanner2Excel}()} function in this package.

The Microsoft Excel \code{.xls(x)} file has on each row for each student in the
following particular order:
\enumerate{
\item the student registration number,
\item the exam version,
\item the responses to the multiple choice exam test items (questions).
}
}
\examples{
## Load the answer key
key <- loadKey(keyFile = paste0(.libPaths()[1], "/wurmc/examples/keyFile.xlsx"))

## Load the student responses from a Microsoft Excel file
responses <- loadResponses(respFile = paste0(.libPaths()[1], "/wurmc/examples/respFile.xlsx"))

## Load the student responses from a comma separated value file as created by
## Formscanner
responses <- loadResponses(respFile = paste0(.libPaths()[1], "/wurmc/examples/respFile.csv"))

}
\seealso{
Other load functions: 
\code{\link{loadKey}()}
}
\author{
Maikel Verouden
}
\concept{load functions}
