% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeFunctions.R
\name{numberCorrect}
\alias{numberCorrect}
\title{Determine the Number of Correct Student Responses}
\usage{
numberCorrect(responses, key)
}
\arguments{
\item{responses}{A data.frame object, as created with the
\code{\link{loadResponses}} function, with in the first column the
student registration numbers of students taking the exam. The second column
specifies the exam version the student filled. Subsequent columns give the
filled in answers to the multiple choice exam test items (questions) as
filled by the students, with a maximum of 40 test items (questions).}

\item{key}{A data.frame object, as create with the \code{\link{loadKey}}
function, with in the first column the exam version and the following
columns giving the correct answers to the multiple choice exam test items
(questions).}
}
\value{
The input \code{responses} data.frame object extended with one column
   indicating the number of multiple choice exam test items (questions)
   answered correctly by a student.
}
\description{
\code{numberCorrect()} checks the student's responses to the multiple choice
exam test items (questions) with the appropriate version key and creates a
column \code{No.correct} indicating the number of test items answered
correctly by a student. The column \code{No.correct} is added after the last
column of the data.frame \code{responses}.
}
\examples{
\dontrun{
responses <- numberCorrect(responses, key)}
}
\seealso{
\code{\link{loadResponses}}, \code{\link{loadKey}}
}
